requirements.txt
aliyun-python-sdk-core-v3==2.8.6
aliyun-python-sdk-dysmsapi==1.0.0
astroid==1.6.1
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.9
Django==2.0.1
django-cors-headers==2.2.0
django-filter==1.1.0
django-mysql==2.2.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
idna==2.6
isort==4.3.4
itchat==1.3.10
lazy-object-proxy==1.3.1
lxml==4.2.1
Markdown==2.6.11
mns==1.1.4
mysqlclient==1.3.12
Pillow==5.0.0
pycrypto==2.6.1
pycryptodome==3.5.1
PyJWT==1.5.3
pypng==0.0.18
PyQRCode==1.2.1
python-pptx==0.6.7
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
wrapt==1.10.11
XlsxWriter==1.0.2
python-dateutil==2.7.2
